Add new file for MainActivity composables
